VisonTransformer/requirement.txt
Requirement

python		3.7.0
cuda  		11.1
pytorch	  	1.9.0
torchvision	0.10.0
pillow
matplotlib	3.3.0
xlwt
tensorflow	2.4.0
numpy		1.19.0
pandas		1.1.0
packaging	20.3